VoterAverage set to double

+++ Components/Models/MovieDetails.cs
@@ -26,7 +26,7 @@
         public string? Tagline { get; set; }
         public string? Title { get; set; }
         public bool Video { get; set; }
-        public int VoteAverage { get; set; }
+        public double VoteAverage { get; set; }
         public int VoteCount { get; set; }
     }
